requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
pymongo==4.5.0  # Atualizado para a versão mais recente
python-decouple==3.8
pytz==2024.2
mongoengine==0.27.0  # Atualizado para a versão mais recente
sqlparse==0.3.1
django-rest-framework-mongoengine==3.4.1